% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{BC}
\alias{BC}
\title{Percentages of blood content (BC)}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{organ}{Human body organs}
  \item{bloodPerc}{Percentage of blood content}
}
}
\source{
\url{https://journals.sagepub.com/doi/pdf/10.1177/ANIB_32_3-4}
}
\description{
A dataset containing the percentages of blood content in human organs.
}
\keyword{internal}
