% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPcoeff.R
\name{calcFup}
\alias{calcFup}
\title{Calculate unbound fraction in plasma}
\usage{
calcFup(logP, type = "total")
}
\arguments{
\item{logP}{Partition coefficient of a molecule between an aqueous and lipophilic phases, usually octanol and water; measurement of lipophilicity}

\item{type}{Type of molecule; can be total, acid, base, or neutral}
}
\value{
Unbound fraction in plasma
}
\description{
Takes in logP and type of molecule and returns the unbound fraction in plasma
}
\details{
Source: https://pubmed.ncbi.nlm.nih.gov/20549836/
}
