% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPcoeff.R
\name{calcKp}
\alias{calcKp}
\title{Calculate partition coefficients for a molecule}
\usage{
calcKp(TCData, logP, pKa = NULL, fup, BP = 1, type = 1, method = "PT")
}
\arguments{
\item{TCData}{Tissue composition data}

\item{logP}{Partition coefficient of a molecule between an aqueous and lipophilic phases, usually octanol and water; measurement of lipophilicity}

\item{pKa}{Negative log of the acid dissociation constant; measurement of the acidic strength of the molecule}

\item{fup}{Unbound fraction of the molecule in plasma}

\item{BP}{Blood:plasma concentration ratio}

\item{type}{Type of the molecule; 1=neutral, 2=monoprotic acid, 3=monoprotic base, 4=diprotic acid, 5=diprotic base, 6=monoprotic acid monoprotic base (acid comes first), 7=triprotic acid, 8=triprotic base, 9=diprotic acid monoprotic base (first two are acid), 10=diprotic base monoprotic acid (first one is acid); default=1}

\item{method}{Prediction method; PT=Poulin and Theil, Berez=Berezhkovskiy, RR=Rodgers and Rowland, Schmitt=Schmitt, pksim=PK-Sim standard; default=PT}
}
\value{
A named list with tissue:plasma partition coefficients
}
\description{
Takes in the drug's physicochemical properties and returns the tissue:plasma partition coefficients
}
