% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{calcKpot}
\alias{calcKpot}
\title{Calculate partition coefficient for the "other" compartment}
\usage{
calcKpot(Kp)
}
\arguments{
\item{Kp}{Named list of tissue:plasma partition coefficients}
}
\value{
Partition coefficient for the "other" compartment
}
\description{
Takes in Kp values for all compartments and returns the partition coefficient of the "other" compartment calculated as the average of the non adipose Kp values
}
\keyword{internal}
