% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{filterDatasets}
\alias{filterDatasets}
\title{Refine datasets based on age and sex}
\usage{
filterDatasets(age, sex)
}
\arguments{
\item{age}{Age}

\item{sex}{Sex}
}
\value{
A named list with the refined datasets based on age and sex
}
\description{
Takes in the target age and sex and returns a list of datasets refined based on these covariates
}
\keyword{internal}
