% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{flow}
\alias{flow}
\title{Percentages of blood flow (flow)}
\format{
A data frame with 21 rows and 2 variables:
\describe{
  \item{organ}{Human body organs}
  \item{flowPerc}{Percentage of blood flow}
}
}
\source{
\url{https://journals.sagepub.com/doi/pdf/10.1177/ANIB_32_3-4}
}
\description{
A dataset containing the percentages of blood flows from cardiac output for human organs.
}
\keyword{internal}
