% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPhys.R
\name{genPop}
\alias{genPop}
\title{Generate population physiological parameters}
\usage{
genPop(
  nSubj,
  minAge,
  maxAge,
  femPerc,
  minBW = NULL,
  maxBW = NULL,
  minHT = NULL,
  maxHT = NULL,
  minBMI = NULL,
  maxBMI = NULL,
  optimize = FALSE,
  addBC = TRUE,
  method = "Willmann"
)
}
\arguments{
\item{nSubj}{Number of subjects}

\item{minAge}{Minimum age in years}

\item{maxAge}{Maximum age in years}

\item{femPerc}{Percentage of females}

\item{minBW}{Minimum body weight in kg}

\item{maxBW}{Maximum body weight in kg}

\item{minHT}{Minimum height in m}

\item{maxHT}{Maximum height in m}

\item{minBMI}{Minimum body mass index in kg/m^2}

\item{maxBMI}{Maximum body mass index in kg/m^2}

\item{optimize}{if TRUE, an optimization step is done; for "Willmann" method}

\item{addBC}{if TRUE, blood content will be added to each organ}

\item{method}{Method to use for generating physiological parameters; choices are "Willmann" and "Huisinga"}
}
\value{
List of named lists with physiological parameters for each individual in the population
}
\description{
Takes in desired population demographics and generates the population physiological parameters
}
