% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{getInterceptSlope}
\alias{getInterceptSlope}
\title{Get slope and intercept for calcBP or calcFup}
\usage{
getInterceptSlope(type = "total", method = NULL, func)
}
\arguments{
\item{type}{Type of molecule; can be total, acid, base, or neutral}

\item{method}{Calculation method; 1=fup-dependent method, 2=logP-dependent method}
}
\value{
A named list with the intercept and slope values
}
\description{
Takes in the method, generic type of molecule, and the function of interest (calcBP or calcFub) and returns the intercept and slope
}
\details{
Source: https://pubmed.ncbi.nlm.nih.gov/20549836/
}
\keyword{internal}
