% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{getKs}
\alias{getKs}
\title{Calculate K_n_l and K_a_pl}
\usage{
getKs(type, pKa, pH, alpha, W, K_n_pl)
}
\arguments{
\item{type}{Type of the molecule; 1=neutral, 2=monoprotic acid, 3=monoprotic base, 4=diprotic acid, 5=diprotic base, 6=monoprotic acid monoprotic base (acid comes first), 7=triprotic acid, 8=triprotic base, 9=diprotic acid monoprotic base (first two are acid), 10=diprotic base monoprotic acid (first one is acid)}

\item{pKa}{Negative log of the acid dissociation constant; measurement of the acidic strength of the molecule}

\item{pH}{pH}

\item{alpha}{Ratio between ditribution coefficient at given pH (D) and that in neutral form (D0)}

\item{W}{W value calculated by the getW function}

\item{K_n_pl}{Neutral phospholipids:water partition coefficient}
}
\value{
A named list with K_n_l and K_a_pl (neutral lipids:water and acidic phospholipids:water partition coefficients, respectively)
}
\description{
Takes in the molecule type, pKa, and pH values and returns W value required by the Schmitt method
}
\keyword{internal}
