% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{getTissueData}
\alias{getTissueData}
\title{Extract information from tissue composition data}
\usage{
getTissueData(dat)
}
\arguments{
\item{dat}{Dataframe containing tissue composition data; columns are: tissue, f_water=water fraction, f_lipids=lipids fraction, f_proteins=protein fraction, f_pl=phospholipids fraction, f_n_l=neutral lipids fraction, f_n_pl=neutral phospholipids fraction, f_a_pl=acidic phospholipids fraction, pH, f_ew=extracellular fraction, f_iw=intracellular fraction, AR=albumin ratio, LR=lipoprotein ratio,  Prediction method; PT=Poulin and Theil, Berez=Berezhkovskiy, RR=Rodgers and Rowland, Schmitt=Schmitt, pksim=PK-Sim standard}
}
\value{
A named list with required components for partition coefficient calculation
}
\description{
Takes in the tissue composition data and returns a named list of the required components for the partition coefficient calculation
}
\keyword{internal}
