% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPhys.R
\name{getVirtPop}
\alias{getVirtPop}
\title{Get the built-in virtual population data}
\usage{
getVirtPop(model = "general")
}
\arguments{
\item{model}{Model type; can be "general" or "mAb"}
}
\value{
List of named lists with physiological parameters for the virtual population
}
\description{
Takes in the model type and returns the built-in virtual population data for 1000 subjects
}
\details{
The covariates used to generate the virtual population are:

Number of subjects: 1000

Age: 20-80 years

Female percentage: 50 percent

Body weight: 50-100 kg

Height: 1.5-1.9 m
}
