% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{getW}
\alias{getW}
\title{Calculate W}
\usage{
getW(type, pKa, pH)
}
\arguments{
\item{type}{Type of the molecule; 1=neutral, 2=monoprotic acid, 3=monoprotic base, 4=diprotic acid, 5=diprotic base, 6=monoprotic acid monoprotic base (acid comes first), 7=triprotic acid, 8=triprotic base, 9=diprotic acid monoprotic base (first two are acid), 10=diprotic base monoprotic acid (first one is acid)}

\item{pKa}{Negative log of the acid dissociation constant; measurement of the acidic strength of the molecule}

\item{pH_IW}{pH of intracellular tissue water}

\item{pH_P}{pH of plasma}

\item{pH_RBC}{pH of red blood cells}
}
\value{
W value required by the Schmitt method
}
\description{
Takes in the molecule type, pKa, and pH values and returns W value required by the Schmitt method
}
\keyword{internal}
