% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{getXYZ}
\alias{getXYZ}
\title{Calculate X, Y, and Z}
\usage{
getXYZ(type, pKa, pH_IW, pH_P, pH_RBC)
}
\arguments{
\item{type}{Type of the molecule; 1=neutral, 2=monoprotic acid, 3=monoprotic base, 4=diprotic acid, 5=diprotic base, 6=monoprotic acid monoprotic base (acid comes first), 7=triprotic acid, 8=triprotic base, 9=diprotic acid monoprotic base (first two are acid), 10=diprotic base monoprotic acid (first one is acid)}

\item{pKa}{Negative log of the acid dissociation constant; measurement of the acidic strength of the molecule}

\item{pH_IW}{pH of intracellular tissue water}

\item{pH_P}{pH of plasma}

\item{pH_RBC}{pH of red blood cells}
}
\value{
A named list with X, Y, and Z values required by the Rodgers and Rowland method
}
\description{
Takes in the molecule type, pKa, and pH values for intracellular water, plasma, and RBCs and returns X, Y, and Z values required by the Rodgers and Rowland method
}
\keyword{internal}
