% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{icrpData}
\alias{icrpData}
\title{ICRP dataset (icrpData)}
\format{
A data frame with 6 rows and 47 variables:
\describe{
  \item{age}{Age, in years}
  \item{bw_m}{Male body weight, in kg}
  \item{bw_f}{Female body weight, in kg}
  \item{ht_m}{Male height, in cm}
  \item{ht_f}{Female height, in cm}
  \item{bsa_m}{Male body surface area, in m^2}
  \item{bsa_f}{Female body surface area, in m^2}
  \item{ad_m}{Male adipose volume, in L}
  \item{ad_f}{Female adipose volume, in L}
  \item{bo_m}{Male bone volume, in L}
  \item{bo_f}{Female bone volume, in L}
  \item{br_m}{Male brain volume, in L}
  \item{br_f}{Female brain volume, in L}
  \item{go_m}{Male gonads volume, in L}
  \item{go_f}{Female gonads volume, in L}
  \item{he_m}{Male heart volume, in L}
  \item{he_f}{Female heart volume, in L}
  \item{ki_m}{Male kidneys volume, in L}
  \item{ki_f}{Female kidneys volume, in L}
  \item{la_int_m}{Male large intestines volume, in L}
  \item{la_int_f}{Female large intestines volume, in L}
  \item{li_m}{Male liver volume, in L}
  \item{li_f}{Female liver volume, in L}
  \item{lu_m}{Male lungs volume, in L}
  \item{lu_f}{Female lungs volume, in L}
  \item{mu_m}{Male muscles volume, in L}
  \item{mu_f}{Female muscles volume, in L}
  \item{pa_m}{Male pancreas volume, in L}
  \item{pa_f}{Female pancreas volume, in L}
  \item{sk_m}{Male skin volume, in L}
  \item{sk_f}{Female skin volume, in L}
  \item{sm_int_m}{Male small intestines volume, in L}
  \item{sm_int_f}{Female small intestines volume, in L}
  \item{sp_m}{Male spleen volume, in L}
  \item{sp_f}{Female spleen volume, in L}
  \item{st_m}{Male stomach volume, in L}
  \item{st_f}{Female stomach volume, in L}
  \item{bl_m}{Male blood volume, in L}
  \item{bl_f}{Female blood volume, in L}
  \item{co_m}{Male cardiac output flow, in L/min}
  \item{co_f}{Female cardiac output flow, in L/min}
  \item{th_m}{Male thymus volume, in L}
  \item{th_f}{Female thymus volume, in L}
  \item{ln_m}{Male lymph nodes volume, in L}
  \item{ln_f}{Female lymph nodes volume, in L}
  \item{ot_m}{Male other organs volume, in L}
  \item{ot_f}{Female other organs volume, in L}
}
}
\source{
\url{https://journals.sagepub.com/doi/pdf/10.1177/ANIB_32_3-4}
}
\description{
Dataset containing physiologic parameters for typical individuals.
}
\keyword{internal}
