% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lnormSD}
\alias{lnormSD}
\title{Geometric standard deviation (lnormSD)}
\format{
A data frame with 4 rows and 2 variables:
\describe{
  \item{organ}{Human body organs}
  \item{geomSD}{Geometric standard deviation}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/17431751/}
}
\description{
A dataset containing the geometric standard deviation for log normally distributed organs.
}
\keyword{internal}
