% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mabShahData}
\alias{mabShahData}
\title{Physiological data for mAb PBPK model (mabShahData)}
\format{
A data frame with 18 rows and 9 variables:
\describe{
  \item{Organ}{Organ}
  \item{Vtot}{Total volume}
  \item{Vplas}{Plasma volume}
  \item{Vbc}{Blood cell volume}
  \item{Vi}{Interstitial volume}
  \item{Ve}{Endosomal volume}
  \item{Vc}{Cellular volume}
  \item{Qplas}{Plasma flow}
  \item{Qbc}{Blood cell flow}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/22143261/}
}
\description{
Physiological data to be used for the monoclonal antibody PBPK model.
}
\keyword{internal}
