% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPhys.R
\docType{data}
\name{nhanesData}
\alias{nhanesData}
\title{Generate individual physiological parameters}
\format{
An object of class \code{data.frame} with 25769 rows and 8 columns.
}
\usage{
nhanesData
}
\arguments{
\item{age}{Age of the individual}

\item{is.male}{if TRUE, individual is male}

\item{bw_targ}{Target body weight}

\item{ht_targ}{Target height}

\item{bmi_targ}{Target body mass index}

\item{optimize}{if TRUE, an optimization step is done}
}
\value{
Named list with physiological parameters for the desired individual
}
\description{
Takes in desired individual demographics and generates the individual physiological parameters
}
\keyword{datasets}
