% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{normSD}
\alias{normSD}
\title{Coefficient of variation (normSD)}
\format{
A data frame with 17 rows and 2 variables:
\describe{
  \item{organ}{Human body organs}
  \item{cv}{Coefficient of variation}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/17431751/}
}
\description{
A dataset containing the coefficient of variation for normally distributed organ volumes.
}
\keyword{internal}
