% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{optimVols}
\alias{optimVols}
\title{Optimize organ volumes}
\usage{
optimVols(optVols)
}
\arguments{
\item{optVols}{Vector of non-optimized organ volumes}
}
\value{
A probability to be minimized by an optimizer
}
\description{
Takes in the a vector of non-optimized organ volumes and returns a probability to be minimized by an optimizer
}
\keyword{internal}
