% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{optimVolsFunc}
\alias{optimVolsFunc}
\title{Optimize organ volumes}
\usage{
optimVolsFunc(optVols, df_opt, df_ad, bw_targ, bw_mean, ht_rel, normOrgan)
}
\arguments{
\item{optVols}{Vector of non-optimized organ volumes}

\item{df_opt}{Dataframe with volumes of organs to be optimized except for adipose}

\item{df_ad}{Dataframe of adipose volume to be optimized}

\item{bw_targ}{Target body weight}

\item{bw_mean}{Mean body weight}

\item{ht_rel}{Relative height ratio}

\item{normOrgan}{Vector of normally distributed organs}
}
\value{
A probability to be minimized by an optimizer
}
\description{
Takes in the a vector of non-optimized organ volumes and returns a probability to be minimized by an optimizer
}
\keyword{internal}
