% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperForTests.R
\name{render_ref}
\alias{render_ref}
\title{Render a ref from a test case row}
\usage{
render_ref(..., .func, out_dir)
}
\description{
This takes a single row, runs the `.func` on it, and dput's the result to a
file. It also returns the same row with a `path` column added, which contains
the path the result was written to.
}
\keyword{**Meant}
\keyword{`pmap_dfr()`**}
\keyword{be}
\keyword{called}
\keyword{internal}
\keyword{to}
\keyword{with}
