% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{sampleIndPars}
\alias{sampleIndPars}
\title{Sample individual parameters}
\usage{
sampleIndPars(
  nSubj,
  minAge,
  maxAge,
  is.male,
  minBW,
  maxBW,
  minHT,
  maxHT,
  minBMI,
  maxBMI,
  optimize = FALSE,
  addBC = TRUE,
  mab = FALSE,
  method = "Willmann"
)
}
\arguments{
\item{nSubj}{Number of subjects}

\item{minAge}{Minimum age in years}

\item{maxAge}{Maximum age in years}

\item{is.male}{if TRUE, individual is male}

\item{minBW}{Minimum body weight in kg}

\item{maxBW}{Maximum body weight in kg}

\item{minHT}{Minimum height in m}

\item{maxHT}{Maximum height in m}

\item{minBMI}{Minimum body mass index in kg/m^2}

\item{maxBMI}{Maximum body mass index in kg/m^2}

\item{optimize}{if TRUE, an optimization step is done for each individual}

\item{addBC}{if TRUE, blood content will be added to each organ}

\item{mab}{if TRUE, mAb model physiological parameters are returned}

\item{method}{Method to use for generating physiological parameters; choices are "Willmann" and "Huisinga"}
}
\value{
A probability to be minimized by an optimizer
}
\description{
Takes in number of subjects, minimum and maximum age, sex, minimum BMI and height and the a dataframe of filtered NHANES data and minimum BMI and height and returns a named list of target body weight, height and BMI
}
\keyword{internal}
