% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{scaleKp}
\alias{scaleKp}
\title{Scale Kp values based on Vss}
\usage{
scaleKp(Kp, Vss, BP, Vt)
}
\arguments{
\item{Kp}{Named list of tissue:plasma partition coefficients}

\item{Vss}{Steady state volume}

\item{BP}{Blood:plasma concentration ratio}

\item{Vt}{Named list of tissue volumes generated by genInd}
}
\value{
A named list of the scaled tissue:plasma partition coefficients
}
\description{
Takes in Kp, Vss, and Vt and returns an error if Vss is provided without Vt
}
\keyword{internal}
