% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{SF}
\alias{SF}
\title{Scaling factors (SF)}
\format{
A data frame with 4 rows and 2 variables:
\describe{
  \item{organ}{Human body organs}
  \item{SF}{Scaling factors}
}
}
\source{
\url{https://github.com/Open-Systems-Pharmacology/OSPSuite.Documentation/wiki/Create-Individual-Algorithm}
}
\description{
A dataset containing the scaling factors for allometric scaling of organs with height.
}
\keyword{internal}
