% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{TCData}
\alias{TCData}
\title{Unified tissue composition data (TCData)}
\format{
A data frame with 13 rows and 13 variables:
\describe{
  \item{tissue}{Tissue}
  \item{f_water}{Fraction of water}
  \item{f_lipids}{Fraction of lipids}
  \item{f_proteins}{Fraction of proteins}
  \item{f_pl}{Fraction of phospholipids}
  \item{f_n_l}{Fraction of neutral lipids}
  \item{f_n_pl}{Fraction of neutral phospholipids}
  \item{f_a_pl}{Fraction of acidic phospholipids}
  \item{pH}{pH}
  \item{f_ew}{Fraction of extracellular water}
  \item{f_iw}{Fraction of intracellular water}
  \item{AR}{Albumin ratio}
  \item{LR}{Lipoprotein ratio}
}
}
\source{
\url{https://dmd.aspetjournals.org/content/48/10/903}
}
\description{
Unified tissue composition data used for the partition coefficient calculation.
}
\keyword{internal}
