% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{test_covRange}
\alias{test_covRange}
\title{Test if the target covariates are out of range}
\usage{
test_covRange(bw_targ, ht_targ, bmi_targ, rangeBW, rangeHT, rangeBMI)
}
\arguments{
\item{bw_targ}{Target body weight in kg}

\item{ht_targ}{Target height in m}

\item{bmi_targ}{Target BMI in kg/m^2}

\item{rangeBW}{Range of BMIs in kg/m^2 for the chosen age and sex}

\item{rangeHT}{Range of heights in m for the chosen age and sex}
}
\value{
An error message if one or more of the covariates are out of range
}
\description{
Takes in the target covariates and corresponding ranges for the chosen age and sex and returns an error if one or more are out of range
}
\keyword{internal}
