% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{test_genIndInput}
\alias{test_genIndInput}
\title{Test if inputs to genInd are compliant}
\usage{
test_genIndInput(age, bw_targ, ht_targ, bmi_targ)
}
\arguments{
\item{bw_targ}{Target body weight in kg}

\item{ht_targ}{Target height in m}

\item{bmi_targ}{Target BMI in kg/m^2}
}
\value{
An error or a warning if inputs to genInd are not compliant
}
\description{
Takes in the target covariates and returns an error or a warning if inputs to genInd are not compliant
}
\keyword{internal}
