% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{test_largeHt}
\alias{test_largeHt}
\title{Test for unrealistic hematocrit value}
\usage{
test_largeHt(Ht)
}
\arguments{
\item{Ht}{Hematocrit value}
}
\value{
An error if Hematocrit > 1
}
\description{
Takes in a named list of Kp values and returns an error if a negative value occurs
}
\keyword{internal}
