% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{test_missingVt}
\alias{test_missingVt}
\title{Test for missing Vt}
\usage{
test_missingVt(Vt)
}
\arguments{
\item{Vt}{Named list of tissue volumes generated by genInd}
}
\value{
An error if Vss is provided without Vt
}
\description{
Takes in Vt and returns an error if Vt is missing
}
\keyword{internal}
