% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{test_pKaTypeMatch}
\alias{test_pKaTypeMatch}
\title{Test matching pKa length and type}
\usage{
test_pKaTypeMatch(type, pKa)
}
\arguments{
\item{type}{Type of the molecule; 1=neutral, 2=monoprotic acid, 3=monoprotic base, 4=diprotic acid, 5=diprotic base, 6=monoprotic acid monoprotic base (acid comes first), 7=triprotic acid, 8=triprotic base, 9=diprotic acid monoprotic base (first two are acid), 10=diprotic base monoprotic acid (first one is acid)}

\item{pKa}{Negative log of the acid dissociation constant; measurement of the acidic strength of the molecule}
}
\value{
An error if pKa length and type don't match
}
\description{
Takes in the molecule type and pKa and throws an error if they don't match
}
\keyword{internal}
