% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{test_realisticInd}
\alias{test_realisticInd}
\title{Test if generated individual is realistic}
\usage{
test_realisticInd(ad)
}
\arguments{
\item{ad}{adipose tissue volume}
}
\value{
An error if unrealistic individual
}
\description{
Takes in the adipose tissue volume and returns an error if value is negative or NA
}
\keyword{internal}
