% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{testCovRange}
\alias{testCovRange}
\title{Get missing covariate}
\usage{
testCovRange(bw_targ, ht_targ, bmi_targ, rangeBW, rangeHT, rangeBMI)

testCovRange(bw_targ, ht_targ, bmi_targ, rangeBW, rangeHT, rangeBMI)
}
\arguments{
\item{bw_targ}{Target body weight}

\item{ht_targ}{Target height}

\item{bmi_targ}{Target BMI}

\item{rangeBW}{Range of BMIs for the chosen age and gender}

\item{rangeHT}{Range of heights for the chosen age and gender}
}
\value{
A named list of target covariates

An error message if one or more of the covariates are out of range
}
\description{
Takes in the target covariates and returns them after calculating the missing one

Takes in the target covariates and corresponding ranges for the chosen age and gender and returns an error if one or more are out of range
}
\keyword{internal}
