% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperForTests.R
\name{vec_to_list}
\alias{vec_to_list}
\title{Maps over a list and wraps any elements that are longer than 1 in list().
This is for converting lists to rows in a tibble with as_tibble()}
\usage{
vec_to_list(.row)
}
\description{
Maps over a list and wraps any elements that are longer than 1 in list().
This is for converting lists to rows in a tibble with as_tibble()
}
\keyword{internal}
