% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{virtPop_gen}
\alias{virtPop_gen}
\title{Virtual population for a general small molecule PBPK model}
\format{
A list of lists with a length of 1000 IDs for the upper level and of 45 for the nested lists
\describe{
  \item{ID}{Subject ID}
  \item{AGE}{Age, in years}
  \item{SEX}{Sex; 1=male and 2=female}
  \item{BW}{Body weight, in kg}
  \item{HT}{Height, in m}
  \item{BMI}{Body mass index, in kg/m^2}
  ...
}
}
\usage{
virtPop_gen
}
\description{
A list of lists with the physiological data for the virtual population of 1000 subjects to be used for a general small molecule PBPK modeling.
}
\details{
Prefix: V=volume, Q=flow

Abbreviation: bo=bone, br=brain, go=gonads, he=heart, ki=kidneys, la_int=large intestines, li=liver, pa=pancreas, sk=skin, sm_int=small intestines, st=stomach, th=thymus, ln=lymph nodes, ot=other, ve=venous, ar=arterial, lu=lungs, mu=muscle, sp=spleen, ad=adipose, ha=hepatic artery
}
\keyword{datasets}
