% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{virtPop_mab}
\alias{virtPop_mab}
\title{Virtual population for the monoclonal antibody PBPK model}
\format{
A list of lists with a length of 1000 IDs for the upper level and of 68 for the nested lists
\describe{
  \item{ID}{Subject ID}
  \item{AGE}{Age, in years}
  \item{SEX}{Sex; 1=male and 2=female}
  \item{BW}{Body weight, in kg}
  \item{HT}{Height, in m}
  \item{BMI}{Body mass index, in kg/m^2}
  ...
}
}
\usage{
virtPop_mab
}
\description{
A list of lists with the physiological data for the virtual population of 1000 subjects to be used for monoclonal antibody PBPK modeling.
}
\details{
Prefix: V=volume, PLQ=flow

Suffix: V=plasma volume, IS=interstitial volume

Abbreviation: LI=large intestines, SI=small intestines
}
\keyword{datasets}
