% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPcoeff.R
\name{calcBP}
\alias{calcBP}
\title{Calculate blood to plasma concentration ratio}
\usage{
calcBP(logP = NULL, fup, method = 1)
}
\arguments{
\item{logP}{Partition coefficient of a molecule between an aqueous and lipophilic phases, usually octanol and water; measurement of lipophilicity}

\item{fup}{Unbound fraction of the molecule in plasma}

\item{method}{BP calculation method; 1=logP-dependent method, 2=fup-dependent method}
}
\value{
Blood to plasma concentration ratio
}
\description{
Takes in logP, fup, and method and returns the blood to plasma concentration ratio
}
\details{
Source: https://pubmed.ncbi.nlm.nih.gov/20549836/
}
