% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPcoeff.R
\name{calcKp}
\alias{calcKp}
\title{Calculate partition coefficients for a molecule}
\usage{
calcKp(
  logP,
  pKa = NULL,
  fup,
  BP = 1,
  type = 1,
  method = "PT",
  Vss = NULL,
  Vt = NULL,
  Kpot = NULL
)
}
\arguments{
\item{logP}{Partition coefficient of a molecule between an aqueous and lipophilic phases, usually octanol and water; measurement of lipophilicity}

\item{pKa}{Negative log of the acid dissociation constant; measurement of the acidic strength of the molecule}

\item{fup}{Unbound fraction of the molecule in plasma}

\item{BP}{Blood:plasma concentration ratio}

\item{type}{Type of the molecule; 1=neutral, 2=monoprotic acid, 3=monoprotic base, 4=diprotic acid, 5=diprotic base, 6=monoprotic acid monoprotic base (acid comes first), 7=triprotic acid, 8=triprotic base, 9=diprotic acid monoprotic base (first two are acid), 10=diprotic base monoprotic acid (first one is acid)}

\item{method}{Prediction method; PT=Poulin and Theil, Berez=Berezhkovskiy, RR=Rodgers and Rowland, Schmitt=Schmitt, pksim=PK-Sim standard}

\item{Vss}{Steady state volume in L}

\item{Vt}{Named list of tissue volumes generated by genInd}

\item{Kpot}{Partition coefficient for the "other" compartment to be used when Vss value is provided. If Kpot is left NULL while Vss value is provided, Kpot will be calculated as the average of the non adipose partition coefficients}
}
\value{
A named list with tissue:plasma partition coefficients
}
\description{
Takes in the molecule's physicochemical properties and returns the tissue:plasma partition coefficients
}
\details{
Sources:

PT: https://pubmed.ncbi.nlm.nih.gov/11782904/

RR: https://pubmed.ncbi.nlm.nih.gov/15858854/ and https://pubmed.ncbi.nlm.nih.gov/16639716/

Berez: https://pubmed.ncbi.nlm.nih.gov/15124219/

Schmitt: https://pubmed.ncbi.nlm.nih.gov/17981004/

pksim: https://www.tandfonline.com/doi/abs/10.1517/17425255.1.1.159
}
