% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPhys.R
\name{genPop}
\alias{genPop}
\title{Generate population physiological parameters}
\usage{
genPop(
  nSubj,
  minAge,
  maxAge,
  femPerc,
  minBW = NULL,
  maxBW = NULL,
  minHT = NULL,
  maxHT = NULL,
  minBMI = NULL,
  maxBMI = NULL,
  optimize = F
)
}
\arguments{
\item{nSubj}{Number of subjects}

\item{minAge}{Minimum age}

\item{maxAge}{Maximum age}

\item{femPerc}{Percentage of females}

\item{minBW}{Minimum body weight}

\item{maxBW}{Maximum body weight}

\item{minHT}{Minimum height}

\item{maxHT}{Maximum height}

\item{minBMI}{Minimum body mass index}

\item{maxBMI}{Maximum body mass index}

\item{optimize}{if TRUE, an optimization step is done for each individual}
}
\value{
List of named lists with physiological parameters for each individual in the population
}
\description{
Takes in desired population demographics and generates the population physiological parameters
}
