% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nhanesData}
\alias{nhanesData}
\title{NHANES dataset (nhanesData)}
\format{
A data frame with 25769 rows and 8 variables:
\describe{
  \item{SEX}{Sex; 1=male and 2=female}
  \item{AGE_YR}{Age in years}
  \item{AGE_MN}{Age in months}
  \item{RACE}{Race; 1=non-Hispanic white, 2=non-Hispanic black, 3=Mexican American, 4=other race, 5=other Hispanic}
  \item{BW}{Body weight, in kg}
  \item{HT}{Height, in cm}
  \item{BMI}{Body mass index, in kg/m^2}
  \item{BSA}{Body surface area, in m^2}
}
}
\source{
\url{https://www.cdc.gov/nchs/nhanes/index.htm}
}
\description{
Dataset containing anthropometric data for individuals 2 years of age and above extracted from NHANES database.
}
\keyword{internal}
