% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{sampleCov}
\alias{sampleCov}
\title{Sample covariates}
\usage{
sampleCov(dat, minBW, maxBW, minHT, maxHT, minBMI, maxBMI)
}
\arguments{
\item{dat}{Dataframe of NHANES data filtered on a specific age and sex}

\item{minBW}{Minimum body weight in kg}

\item{maxBW}{Maximum body weight in kg}

\item{minHT}{Minimum height in m}

\item{maxHT}{Maximum height in m}

\item{minBMI}{Minimum body mass index in kg/m^2}

\item{maxBMI}{Maximum body mass index in kg/m^2}
}
\value{
A named list of target body weight, height and BMI
}
\description{
Takes in the a dataframe of filtered NHANES data and minimum BMI and height and returns a named list of target body weight, height and BMI
}
\keyword{internal}
