% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{sampleCov}
\alias{sampleCov}
\title{Sample covariates}
\usage{
sampleCov(dat, minBW, maxBW, minHT, maxHT, minBMI, maxBMI)
}
\arguments{
\item{dat}{Dataframe of NHANES data filtered on a specific age and sex}

\item{minHT}{Minimum height for the chosen age and sex}

\item{minBMI}{Minimum BMI for the chosen age and sex}
}
\value{
A probability to be minimized by an optimizer
}
\description{
Takes in the a dataframe of filtered NHANES data and minimum BMI and height and returns a named list of target body weight, height and BMI
}
\keyword{internal}
