% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{sampleIndPars}
\alias{sampleIndPars}
\title{Sample individual parameters}
\usage{
sampleIndPars(nSubj, minAge, maxAge, is.male, minBMI, minHT, optimize = FALSE)
}
\arguments{
\item{nSubj}{Number of subjects}

\item{minAge}{Minimum age}

\item{maxAge}{Maximum age}

\item{is.male}{if TRUE, individual is male}

\item{minBMI}{Minimum BMI for the chosen age and sex}

\item{minHT}{Minimum height for the chosen age and sex}

\item{optimize}{if TRUE, an optimization step is done for each individual}
}
\value{
A probability to be minimized by an optimizer
}
\description{
Takes in number of subjects, minimum and maximum age, sex, minimum BMI and height and the a dataframe of filtered NHANES data and minimum BMI and height and returns a named list of target body weight, height and BMI
}
\keyword{internal}
