% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{sampleIndPars}
\alias{sampleIndPars}
\title{Sample individual parameters}
\usage{
sampleIndPars(
  nSubj,
  minAge,
  maxAge,
  is.male,
  minBW,
  maxBW,
  minHT,
  maxHT,
  minBMI,
  maxBMI,
  optimize = FALSE,
  addBC = TRUE
)
}
\arguments{
\item{nSubj}{Number of subjects}

\item{minAge}{Minimum age}

\item{maxAge}{Maximum age}

\item{is.male}{if TRUE, individual is male}

\item{minBW}{Minimum body weight}

\item{maxBW}{Maximum body weight}

\item{minHT}{Minimum height}

\item{maxHT}{Maximum height}

\item{minBMI}{Minimum body mass index}

\item{maxBMI}{Maximum body mass index}

\item{optimize}{if TRUE, an optimization step is done for each individual}

\item{addBC}{if TRUE, blood content will be added to each organ}
}
\value{
A probability to be minimized by an optimizer
}
\description{
Takes in number of subjects, minimum and maximum age, sex, minimum BMI and height and the a dataframe of filtered NHANES data and minimum BMI and height and returns a named list of target body weight, height and BMI
}
\keyword{internal}
