% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{test_genPopInput}
\alias{test_genPopInput}
\title{Test if inputs to genPop are compliant}
\usage{
test_genPopInput(minBW, maxBW, minHT, maxHT, minBMI, maxBMI)
}
\arguments{
\item{minBW}{Minimum body weight}

\item{maxBW}{Maximum body weight}

\item{minHT}{Minimum height}

\item{maxHT}{Maximum height}

\item{minBMI}{Minimum body mass index}

\item{maxBMI}{Maximum body mass index}
}
\value{
An error or a warning if inputs to genPop are not compliant
}
\description{
Takes in the target covariates and returns an error or a warning if inputs to genPop are not compliant
}
\keyword{internal}
