% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{test_NegativeKps}
\alias{test_NegativeKps}
\title{Test for negative Kp values}
\usage{
test_NegativeKps(Kps)
}
\arguments{
\item{Kps}{Named list with Kp values}
}
\value{
An error if one or more Kp values is < 0
}
\description{
Takes in a named list of Kp values and returns an error if a negative value occurs
}
\keyword{internal}
